This version is designed to be efficient with memory.  All static strings are moved to program memory space (flash) instead of SRAM.  This is because this code is designed to run on an arduino with an SD card reader, and there's a load of new code required to be using that.  Still some program space left, but only a couple of hundred bytes most of the time.
Note that this code is also set up to use a slightly modified version of the AFMotor library, with some switch-a-roos in it to change the pins that are used to drive the motorshield.  This is because the SD card is hard-wired to use the SPI buss which is on pin 10, 11, 12 and 13, and the motorshield already uses 11 and 12.  So physically, those four pins are no longer directly connected to the motorshield, and instead we are piping instructions to pin 9 instead of pin 11, and to pin A0 instead of pin 12.  I'm using simple jumpers leads on the board itself to patch them back into the proper places.

I will fork this code into a non-SD card reading version too. At some point.

git-svn-id: http://polargraph.googlecode.com/svn/embedded/trunk@11 5ebda15a-75ab-8a5e-e99a-b4cc1c219dce